backend/requirements.txt
fastapi
uvicorn[standard]
pyserial
pyocd
python-multipart
